+++ packages/exo/test/exo-wobbly-point.test.js
@@ -9,6 +9,7 @@
 /* eslint-disable max-classes-per-file */
 import test from '@endo/ses-ava/test.js';
 
+import harden from '@endo/harden';
 import { getMethodNames } from '@endo/eventual-send/utils.js';
 import { passStyleOf, Far, GET_METHOD_NAMES } from '@endo/pass-style';
 import { M, getNamedMethodGuards } from '@endo/patterns';