+++ packages/import-bundle/test/compartment-wrapper.test.js
@@ -1,5 +1,6 @@
 import test from '@endo/ses-ava/test.js';
 
+import harden from '@endo/harden';
 import { wrapInescapableCompartment } from '../src/compartment-wrapper.js';
 
 const createChild = `() => new Compartment({console})`;
@@ -39,7 +40,9 @@ function check(t, c, n) {
 // `dependencies`), to allow a more thorough test, which is currently
 // only really exercised by the agoric-sdk/swingset-liveslots test
 
-test('wrap', t => {
+// The inescapable compartment scenario does not and is not expected to work
+// without lockdown.
+(Object.isFrozen(Object) ? test : test.skip)('wrap', t => {
   const inescapableTransforms = [];
   const inescapableGlobalProperties = {
     WeakMap: 'replaced',