+++ packages/netstring/reader.js
@@ -1,5 +1,6 @@
 // @ts-check
-/// <reference types="ses"/>
+
+import harden from '@endo/harden';
 
 const COLON = ':'.charCodeAt(0);
 const COMMA = ','.charCodeAt(0);